+++ packages/fighting-design/input/src/props.ts
@@ -29,9 +29,8 @@ export const Props = {
    * 尺寸
    *
    * @values large middle small mini
-   * @default null
    */
-  size: setStringProp<FightingSize>(null, (val: FightingSize): boolean => {
+  size: setStringProp<FightingSize>(undefined, (val: FightingSize): boolean => {
     return FIGHTING_SIZE.includes(val)
   }),
   /**
@@ -92,14 +91,8 @@ export const Props = {
   height: setStringNumberProp(),
   /** 是否可搜索 */
   search: setBooleanProp(),
-  /** 自定义背景色 */
-  background: setStringProp(),
   /** 自定义文字颜色 */
   textColor: setStringProp(),
-  /** 自定义获取焦点背景色 */
-  activeBackground: setStringProp(),
-  /** 自定义获取焦点边框颜色 */
-  activeBorderColor: setStringProp(),
   /**
    * 是否只读
    *